requirements.txt
flask==3.0.0
pillow==10.1.0
gunicorn==21.2.0
flask-cors==4.0.0
